Fix for x64 about using %d on size_t

+++ vs2019/linux-simple-print/simple-print.cpp
@@ -2,7 +2,7 @@
 
 int main(int argc, char *argv[])
 {
-	printf("sizeof(void*) = %d\n", sizeof(void*));
+	printf("sizeof(void*) = %d\n", (int)sizeof(void*));
 	printf("argc = %d\n", argc);
 	printf("argv[0] = %s\n", argv[0]);
